+++ usecase/usecase_suite.go
@@ -2,8 +2,6 @@ package usecase
 
 import (
 	lib "allora_offchain_node/lib"
-
-	"github.com/rs/zerolog/log"
 )
 
 type UseCaseSuite struct {
@@ -16,7 +14,6 @@ func NewUseCaseSuite(userConfig lib.UserConfig) (*UseCaseSuite, error) {
 	userConfig.ValidateConfigAdapters()
 	nodeConfig, err := userConfig.GenerateNodeConfig()
 	if err != nil {
-		log.Error().Err(err).Msg("Failed to initialize allora client")
 		return nil, err
 	}
 	return &UseCaseSuite{Node: *nodeConfig}, nil