+++ lib/repo_tx_registration.go
@@ -47,7 +47,11 @@ func (node *NodeConfig) RegisterWorkerIdempotently(config WorkerConfig) bool {
 	}
 	res, err := node.SendDataWithRetry(ctx, msg, "Register worker node")
 	if err != nil {
-		log.Error().Err(err).Uint64("topic", config.TopicId).Str("txHash", res.TxHash).Msg("Could not register the worker node with the Allora blockchain")
+		txHash := ""
+		if res != nil {
+			txHash = res.TxHash
+		}
+		log.Error().Err(err).Uint64("topic", config.TopicId).Str("txHash", txHash).Msg("Could not register the worker node with the Allora blockchain")
 		return false
 	}
 
@@ -93,7 +97,11 @@ func (node *NodeConfig) RegisterAndStakeReputerIdempotently(config ReputerConfig
 		}
 		res, err := node.SendDataWithRetry(ctx, msgRegister, "Register reputer node")
 		if err != nil {
-			log.Error().Err(err).Uint64("topic", config.TopicId).Str("txHash", res.TxHash).Msg("Could not register the reputer node with the Allora blockchain")
+			txHash := ""
+			if res != nil {
+				txHash = res.TxHash
+			}
+			log.Error().Err(err).Uint64("topic", config.TopicId).Str("txHash", txHash).Msg("Could not register the reputer node with the Allora blockchain")
 			return false
 		}
 
@@ -107,7 +115,7 @@ func (node *NodeConfig) RegisterAndStakeReputerIdempotently(config ReputerConfig
 	}
 	minStake := cosmossdk_io_math.NewInt(config.MinStake)
 	if minStake.LTE(stake) {
-		log.Error().Msg("Reputer stake below minimum stake, skipping.")
+		log.Info().Msg("Reputer stake above minimum requested stake, skipping adding stake.")
 		return true
 	}
 
@@ -118,7 +126,11 @@ func (node *NodeConfig) RegisterAndStakeReputerIdempotently(config ReputerConfig
 	}
 	res, err := node.SendDataWithRetry(ctx, msgAddStake, "Add reputer stake")
 	if err != nil {
-		log.Error().Err(err).Uint64("topic", config.TopicId).Str("txHash", res.TxHash).Msg("Could not stake the reputer node with the Allora blockchain in specified topic")
+		txHash := ""
+		if res != nil {
+			txHash = res.TxHash
+		}
+		log.Error().Err(err).Uint64("topic", config.TopicId).Str("txHash", txHash).Msg("Could not stake the reputer node with the Allora blockchain in specified topic")
 		return false
 	}
 	return true