+++ CHANGELOG.md
@@ -46,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * [#63](https://github.com/allora-network/allora-offchain-node/pull/63) Loss Function Library support.
+* [#66](https://github.com/allora-network/allora-offchain-node/pull/66) Smart worker detection of open window + randomness
 
 ### Removed