+++ CHANGELOG.md
@@ -47,6 +47,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * [#63](https://github.com/allora-network/allora-offchain-node/pull/63) Loss Function Library support.
 * [#65](https://github.com/allora-network/allora-offchain-node/pull/65) Introduced different retry delays for account sequence.
+* [#68](https://github.com/allora-network/allora-offchain-node/pull/68) Logging configuration
 
 ### Removed